Use cell offset in dragging
Dragging a CellItem can start from on top of a different cell than the one where the item is placed at.
For example when the item spans multiple cells and the dragging
operation starts from the third cell to the right of where the cell the
item is placed in.

In these cases we need to calculate an offset of how many cells to the
right the dragging started and use that offset when moving the item.
This prevents the items abruptly snapping to different cells when
dragging them from their right sides.